refactor: FestivalList 컴포넌트 삭제
