+++ apps/audience/src/pages/my-events/my-events.tsx
@@ -1,30 +1,34 @@
 import { useQuery } from '@tanstack/react-query';
+import { useNavigate } from 'react-router';
 
 import { EmptyView } from '@amp/ads-ui';
 import { Loading } from '@amp/compositions';
 
-import FestivalList from '@widgets/my-events/festival-list';
+import FestivalCard from '@widgets/home/components/festival-card/festival-card';
 
-import { MY_EVENTS_QUERY_OPTIONS } from '@features/my-events/query';
 import { MY_PAGE_QUERY_OPTIONS } from '@features/mypage/apis/query';
+
+import { ROUTE } from '@shared/constants/path';
 
 import * as styles from './my-events.css';
 
 const MyEventsPage = () => {
-  const { data: viewedData } = useQuery(
+  const navigate = useNavigate();
+  const { data: viewedData, isPending } = useQuery(
     MY_PAGE_QUERY_OPTIONS.VIEWED_FESTIVALS(),
-  );
-  const { data, isPending } = useQuery(
-    MY_EVENTS_QUERY_OPTIONS.LIST({ page: 0, size: 20 }),
   );
 
   const festivals = viewedData?.festivals ?? [];
+
+  const handleCardClick = (festivalId: number) => {
+    navigate(ROUTE.noticeList(festivalId));
+  };
 
   if (isPending) {
     return <Loading />;
   }
 
-  if (!data || festivals.length === 0) {
+  if (!viewedData || festivals.length === 0) {
     return (
       <section className={styles.page}>
         <div className={styles.empty}>
@@ -37,7 +41,16 @@ const MyEventsPage = () => {
   return (
     <section className={styles.page}>
       <div className={styles.list}>
-        <FestivalList festivals={festivals} />
+        {festivals.map((festival) => {
+          return (
+            <FestivalCard
+              key={festival.festivalId}
+              festival={festival}
+              showWishList={false}
+              onCardClick={handleCardClick}
+            />
+          );
+        })}
       </div>
     </section>
   );
