--- apps/audience/src/shared/types/my-events-response.ts
@@ -1,6 +1,0 @@
-import type { Festival, PaginationResponse } from './festival';
-
-export interface MyEventsResponse {
-  festivals: Festival[];
-  pagination: PaginationResponse;
-}
